// src/features/TodolistsList/api/todolistsApi.types.ts
import {TaskPriorities, TaskStatuses} from "../../../common/enums/index";

export type TodolistType = {
  id: string;
  title: string;
  addedDate: string;
  order: number;
};

export type LoginParamsType = {
    email: string;
    password: string;
    rememberMe: boolean;
    captcha?: string;
};

export type ResponseType<D = {}> = {
    resultCode: number;
    messages: Array<string>;
    data: D;
};


export type TaskType = {
  description: string;
  title: string;
  status: TaskStatuses;
  priority: TaskPriorities;
  startDate: string;
  deadline: string;
  id: string;
  todoListId: string;
  order: number;
  addedDate: string;
};
export type UpdateTaskModelType = {
  title: string;
  description: string;
  status: TaskStatuses;
  priority: TaskPriorities;
  startDate: string;
  deadline: string;
};
export type GetTasksResponse = {
  error: string | null;
  totalCount: number;
  items: TaskType[];
};